Implicit cast fixes for noise.

+++ libraries/stdlib/genglsl/lib/mx_noise.glsl
@@ -161,8 +161,8 @@ float mx_fade(float t)
 uint mx_hash_int(int x, int y)
 {
     uint a, b, c;
-    uint len = 2;
-    a = b = c = 0xdeadbeef + (len << 2) + 13;
+    uint len = 2u;
+    a = b = c = uint(0xdeadbeef) + (len << 2u) + 13u;
     a += x;
     b += y;
     c = mx_bjfinal(a, b, c);
@@ -172,8 +172,8 @@ uint mx_hash_int(int x, int y)
 uint mx_hash_int(int x, int y, int z)
 {
     uint a, b, c;
-    uint len = 3;
-    a = b = c = 0xdeadbeef + (len << 2) + 13;
+    uint len = 3u;
+    a = b = c = uint(0xdeadbeef) + (len << 2u) + 13u;
     a += x;
     b += y;
     c += z;
